src/requirements.txt
aiofiles==24.1.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
greenlet==3.1.1
html5tagger==1.3.0
httptools==0.6.4
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
sanic==24.12.0
sanic-routing==23.12.0
SQLAlchemy==1.4.54
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.21.0
websockets==14.2
